+++ backend/de.metas.business/src/main/java-legacy/org/compiere/model/MPayment.java
@@ -22,6 +22,8 @@
 package org.compiere.model;
 
 import de.metas.allocation.api.IAllocationDAO;
+import de.metas.banking.BankId;
+import de.metas.banking.api.BankRepository;
 import de.metas.bpartner.BPartnerBankAccountId;
 import de.metas.bpartner.service.BPartnerCreditLimitRepository;
 import de.metas.bpartner.service.BPartnerStats;
@@ -53,8 +55,7 @@ import de.metas.payment.api.impl.PaymentBL;
 import de.metas.util.Check;
 import de.metas.util.Services;
 import de.metas.util.StringUtils;
-import org.adempiere.bank.BankId;
-import org.adempiere.bank.BankRepository;
+
 import org.adempiere.exceptions.AdempiereException;
 import org.adempiere.exceptions.FillMandatoryException;
 import org.adempiere.service.ClientId;
