+++ backend/de.metas.business/src/main/java/de/metas/banking/api/BankRepository.java
@@ -20,8 +20,9 @@
  * #L%
  */
 
-package org.adempiere.bank;
+package de.metas.banking.api;
 
+import de.metas.banking.BankId;
 import de.metas.bpartner.BPartnerBankAccountId;
 import de.metas.cache.CCache;
 import de.metas.util.Services;
